gjmöhmgnöh

+++ dental-clinic-system/backend/api/appointment/appointment.go
@@ -2,6 +2,8 @@ package appointment
 
 import (
 	"dental-clinic-system/application/appointmentService"
+	"dental-clinic-system/application/userService"
+	"dental-clinic-system/helpers"
 	"dental-clinic-system/models"
 	"encoding/json"
 	"github.com/gorilla/mux"
@@ -10,16 +12,31 @@ import (
 	"net/http"
 )
 
-func NewAppointmentHandlerController(service appointmentService.AppointmentService) *AppointmentHandler {
-	return &AppointmentHandler{appointmentService: service}
+func NewAppointmentHandlerController(appointmentService appointmentService.AppointmentService, userService userService.UserService) *AppointmentHandler {
+
+	return &AppointmentHandler{
+		appointmentService: appointmentService,
+		userService:        userService,
+	}
 }
 
 type AppointmentHandler struct {
 	appointmentService appointmentService.AppointmentService
+	userService        userService.UserService
 }
 
 func (h *AppointmentHandler) GetAppointments(w http.ResponseWriter, r *http.Request) {
-	appointments, err := h.appointmentService.GetAppointments()
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusNotFound)
+		return
+	}
+
+	appointments, err := h.appointmentService.GetAppointments(user.ClinicID)
+
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusNotFound)
 		return
@@ -38,11 +55,26 @@ func (h *AppointmentHandler) GetAppointment(w http.ResponseWriter, r *http.Reque
 		http.Error(w, "Invalid appointmentRepository ID", http.StatusBadRequest)
 		return
 	}
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, "User not found", http.StatusUnauthorized)
+		return
+	}
+
 	appointment, err := h.appointmentService.GetAppointment(uint(id))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusNotFound)
 		return
 	}
+
+	if user.ClinicID != appointment.ClinicID {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	err = json.NewEncoder(w).Encode(appointment)
 	if err != nil {
 		return
@@ -55,7 +87,18 @@ func (h *AppointmentHandler) CreateAppointment(w http.ResponseWriter, r *http.Re
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-	appointment, err := h.appointmentService.CreateAppointment(appointment)
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, "User not found", http.StatusUnauthorized)
+		return
+	}
+
+	appointment.ClinicID = user.ClinicID
+
+	appointment, err = h.appointmentService.CreateAppointment(appointment)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -72,7 +115,27 @@ func (h *AppointmentHandler) UpdateAppointment(w http.ResponseWriter, r *http.Re
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-	appointment, err := h.appointmentService.UpdateAppointment(appointment)
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, "User not found", http.StatusUnauthorized)
+		return
+	}
+
+	GetAppointment, err := h.appointmentService.GetAppointment(appointment.ID)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusNotFound)
+		return
+	}
+
+	if user.ClinicID != GetAppointment.ClinicID {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
+	appointment, err = h.appointmentService.UpdateAppointment(appointment)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -91,6 +154,26 @@ func (h *AppointmentHandler) DeleteAppointment(w http.ResponseWriter, r *http.Re
 		http.Error(w, "Invalid appointmentRepository ID", http.StatusBadRequest)
 		return
 	}
+
+	claims := helpers.TokenEmailHelper(r)
+	user, err := h.userService.GetUserByEmail(claims.Email)
+
+	if err != nil {
+		http.Error(w, "User not found", http.StatusUnauthorized)
+		return
+	}
+
+	GetAppointment, err := h.appointmentService.GetAppointment(uint(id))
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusNotFound)
+		return
+	}
+
+	if user.ClinicID != GetAppointment.ClinicID {
+		http.Error(w, "Unauthorized", http.StatusUnauthorized)
+		return
+	}
+
 	err = h.appointmentService.DeleteAppointment(uint(id))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
